Initial script commit
Initial commit, generates tables similar to the ACPI SLIT table.
Examples:

dual-socket, 6-core system (HT disabled):

  |  0  1  2  3  4  5  6  7  8  9 10 11
--+------------------------------------
 0| 10  9  9  9  9  9  7  7  7  7  7  7
 1|  9 10  9  9  9  9  7  7  7  7  7  7
 2|  9  9 10  9  9  9  7  7  7  7  7  7
 3|  9  9  9 10  9  9  7  7  7  7  7  7
 4|  9  9  9  9 10  9  7  7  7  7  7  7
 5|  9  9  9  9  9 10  7  7  7  7  7  7
 6|  7  7  7  7  7  7 10  9  9  9  9  9
 7|  7  7  7  7  7  7  9 10  9  9  9  9
 8|  7  7  7  7  7  7  9  9 10  9  9  9
 9|  7  7  7  7  7  7  9  9  9 10  9  9
10|  7  7  7  7  7  7  9  9  9  9 10  9
11|  7  7  7  7  7  7  9  9  9  9  9 10

The diagonal is always 10, aka 100% for CPU local latencies.  The
remaining CPUs are normalized to this.  Cores on the same socket
generall show 9 or 8, ie. 80% or 90%.  Cores on remote sockets are
expected to be noticeably lower, 70% here.

single-socket, 4-core (HT enabled):

  |  0  1  2  3  4  5  6  7
--+------------------------
 0| 10  9  9  9  7  9  9  9
 1|  9 10  9  8  9  7  9  9
 2|  8  9 10  8  9  9  6  8
 3|  8  8  8 10  8  8  8  6
 4|  7  9  9  9 10  9  9  9
 5|  9  7  9  9  9 10  9  9
 6|  8  9  6  9  8  9 10  8
 7|  8  9  9  7  8  8  9 10

Here we can see the same general pattern for a single node system,
but we can verify the thread-pair of the core, which has a performace
dip versus the others.  For instance, the intersection of CPUs 0 and
4 (or 4 and 0) shows 70%, showing the performance impact of running
client and server on separate threads of the same core.

Generally this doesn't show anything that isn't available via lscpu
or numactl, but those tools rely on firmware tables and just
regurgitates how the kernel has interpretted this.  This tool is
meant to actually validate through testing that they are correct.

To run, simply download the latest netperf
(ftp://ftp.netperf.org/netperf/), configure, make, make install.
This will install to /usr/local, which is what this script expects
(update if necessary).  The taskset and bc utilities are the only
other requirements, both of which should be pretty common.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>